images: Redo all images.
We now hold 2 set of product images.

One in legacy-images for images in jpg format for use by platforms who
don't support webp in 2024

And another in live-images that have webp images without a background.

All images now have a resolution of 700x700.

The old images/ directory is still there until search engines switch off
ot it.

A thing to do in the future is to ship minifed images to use for product
feeds.